requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlmodel==0.0.19
python-dotenv==1.0.1
python-telegram-bot==20.7
httpx==0.25.2
google-generativeai
Pillow
pillow-heif
